Somewhat crude filtering added; also ability to view batches from other users

+++ batches_oauth.php
@@ -11,6 +11,7 @@ $action = get_request ( 'action' , '' ) ;
 $batch_id = get_request ( 'id' , '' ) ;
 $page = intval(get_request ( 'page', '1' ));
 $limit = intval(get_request ( 'limit', '100' ));
+$filter = get_request ('filter', 'All') ;
 
 if ($action == 'authorize') {
 	$oauth->doAuthorizationRedirect($oauth_url_prefix . 'batches_oauth.php');
@@ -99,53 +100,99 @@ print "<hr>";
 
 $owner = $oauth->userinfo->name;
 
+$viewall = get_request ('viewall', '');
+
 if ( $batch_id  == '') {
-	$batch_count = Batch::batches_count($db_conn, $owner);
-	$max_page = ($batch_count - 1)/$limit + 1;
-	print "<div>$batch_count total batches</div>";
-	print "<div>";
-	print "<a href='?page=1'>latest</a> | <a href='?page=$max_page'>earliest</a>";
-	$prev_page = $page - 1;
-	$next_page = $page + 1;
-	if ($prev_page > 0) {
-		print " | <a href='?page=$prev_page'>newer</a>";
+	print '<div>';
+	if ($viewall == '') {
+		print '<b>My batches</b> - <a href="?viewall=true">View batches from all users</a>';
 	} else {
-		print "| newer";
+		print '<b>All batches</b> - <a href="?viewall=">View just my batches</a>';
 	}
-	if ($next_page <= $max_page) {
-		print " | <a href='?page=$next_page'>older</a>";
-	} else {
-		print "| older";
+	print '</div>';
+	$filter_options = ['All', 'Queued', 'Running', 'Stopped', 'Errors', 'Done'];
+	print "<form method='get' class='form form-inline'>";
+	print "<input type='hidden' name='page' value='$page' />";
+	print "<input type='hidden' name='limit' value='$limit' />";
+	print "<input type='hidden' name='viewall' value='$viewall' />";
+	print "<select name='filter'>" ;
+	foreach ($filter_options AS $filter_option) {
+		print "<option value='$filter_option'" ;
+		if ($filter == $filter_option) print ' selected' ;
+		print ">$filter_option</option>" ;
 	}
-	print "</div>";
-	print "<form method='post' class='form'>" ;
-	print "<input type='hidden' name='action' value='delete' />";
-?>
+	print "</select>&nbsp; <input type='submit' class='btn btn-primary' name='doit' value='Filter batches' /></form>";
 
-<div>
-<a href='#' onclick='$($(this).parents("form")).find("input[type=checkbox]").prop("checked",true);return false'>Check all</a> | 
-<a href='#' onclick='$($(this).parents("form")).find("input[type=checkbox]").prop("checked",false);return false'>Uncheck all</a>
+	$batch_list = [];
+	if ($viewall == '') {
+		$batch_count = Batch::batches_count($db_conn, $owner);
+		$max_page = ($batch_count - 1)/$limit + 1;
+		print "<div>$batch_count total batches</div>";
+		print "<div>";
+		print "<a href='?page=1'>latest</a> | <a href='?page=$max_page'>earliest</a>";
+		$prev_page = $page - 1;
+		$next_page = $page + 1;
+		if ($prev_page > 0) {
+			print " | <a href='?page=$prev_page'>newer</a>";
+		} else {
+			print "| newer";
+		}
+		if ($next_page <= $max_page) {
+			print " | <a href='?page=$next_page'>older</a>";
+		} else {
+			print "| older";
+		}
+		print "</div>";
+		print "<form method='post' class='form'>" ;
+		print "<input type='hidden' name='action' value='delete' />";
+		print '<div>
+<a href="#" onclick=\'$($(this).parents("form")).find("input[type=checkbox]").prop("checked",true);return false\'>Check all</a> | 
+<a href="#" onclick=\'$($(this).parents("form")).find("input[type=checkbox]").prop("checked",false);return false\'>Uncheck all</a>
 </div>
+';
+		$batch_list = Batch::batches_for_owner($db_conn, $owner, $limit, $page);
+		$delete_count = 0;
+	} else {
+		$batch_list = Batch::all_batches($db_conn, $limit, $page);
+	}
 
-<?PHP
-
-	$batch_list = Batch::batches_for_owner($db_conn, $owner, $limit, $page);
-	$delete_count = 0;
-
-	print "<table class='table table-striped table-condensed'><tr><th></th><th>Batch ID</th><th>Start time</th><th>Counts</th><th>Still processing?</th><th></th></tr>";
+	print "<table class='table table-striped table-condensed'>";
+	print "<tr><th>";
+	if ($viewall != '') print "Owner";
+	print "</th><th>Batch ID</th><th>Start time</th><th>Counts</th><th>Still processing?</th><th></th></tr>";
 	foreach ($batch_list AS $batch) {
 		$id = $batch->batch_id;
 		$display_counts = array();
 		$is_running = $batch->is_running();
+		$is_queued = $batch->queued;
 		$has_ready = $batch->has_ready();
 		$has_error = $batch->has_error();
+		if ($filter != 'All') {
+			if ($is_running) {
+				if ($filter != 'Running') continue;
+			} else if ($is_queued) {
+				if ($filter != 'Queued') continue;
+			} else {
+				if ($has_ready) {
+					if ($filter != 'Stopped') continue;
+				} else if ($has_error) {
+					if ($filter != 'Errors') continue;
+				} else {
+					if ($filter != 'Done') continue;
+				}
+			}
+		}
 		foreach ($batch->counts AS $status => $count) {
 			$display_counts[] = "$status($count)";
 		}
 		print "<tr><td>";
-		if (!$is_running ) {
-			$delete_count += 1;
-			print "<input type='checkbox' name='deletions[$id]' value='$id'/></td>" ;
+		if ($viewall == '') {
+			if (!$is_running ) {
+				$delete_count += 1;
+				print "<input type='checkbox' name='deletions[$id]' value='$id'/></td>" ;
+			}
+		} else {
+			print $batch->owner;
 		}
 		print "</td>";
 		print "<td><a href='?id=$id'>$id</a></td>";
@@ -154,32 +201,40 @@ if ( $batch_id  == '') {
 
 		if ( $is_running ) {
 			print "<td>Yes</td>";
-			print "<td><a href='?action=stop&batch_id=$id'>Stop batch?</a></td>";
+			if ($viewall == '') {
+				print "<td><a href='?action=stop&batch_id=$id'>Stop batch?</a></td>";
+			}
 		} else if ($batch->queued) {
 			print "<td>Queued</td>";
-			print "<td><a href='?action=remove_from_queue&batch_id=$id'>Remove from queue?</a></td>";
+			if ($viewall == '') {
+				print "<td><a href='?action=remove_from_queue&batch_id=$id'>Remove from queue?</a></td>";
+			}
 		} else {
 			print "<td>No</td>";
-			if ($has_ready) {
-				print "<td><a href='?action=restart&batch_id=$id'>Restart batch?</a></td>";
-			} else if ($has_error) {
-				print "<td><a href='?action=reset&batch_id=$id'>Reset errors?</a></td>";
-			} else {
-				print "<td><a href='?action=delete&batch_id=$id'>Delete batch?</a></td>";
+			if ($viewall == '') {
+				if ($has_ready) {
+					print "<td><a href='?action=restart&batch_id=$id'>Restart batch?</a></td>";
+				} else if ($has_error) {
+					print "<td><a href='?action=reset&batch_id=$id'>Reset errors?</a></td>";
+				} else {
+					print "<td><a href='?action=delete&batch_id=$id'>Delete batch?</a></td>";
+				}
 			}
 		}
 		
 		print "</tr>";
 	}
 	print "</table>";
-	if ($delete_count > 0) {
-		print "<div style='margin:20px'><input type='submit' name='doit' value='Delete selected batches' class='btn btn-primary' /> </div>";
+	if ($viewall == '') {
+		if ($delete_count > 0) {
+			print "<div style='margin:20px'><input type='submit' name='doit' value='Delete selected batches' class='btn btn-primary' /> </div>";
+		}
+		print "</form>" ;
 	}
-	print "</form>" ;
 } else {
 	$batch = new Batch($batch_id);
 	$batch->load($db_conn);
-	print "<h3>Batch $batch_id started " . $batch->start_date . "</h3>\n";
+	print "<h3>Batch $batch_id for " . $batch->owner . " started " . $batch->start_date . "</h3>\n";
 	$reload_js = '<script type="text/javascript">
 $(document).ready ( function () {
 	setTimeout(function() { window.location.reload() }, 10000);
@@ -192,16 +247,20 @@ $(document).ready ( function () {
 		print($reload_js);
 	} else if ($batch->queued) {
 		print("Queued (will run after earlier batches complete) ");
-		print "<a href='?id=$batch_id&action=remove_from_queue&batch_id=$batch_id'>Remove from queue?</a>";
+		if ($owner == $batch->owner) {
+			print "<a href='?id=$batch_id&action=remove_from_queue&batch_id=$batch_id'>Remove from queue?</a>";
+		}
 		print($reload_js);
 	} else {
 		print("Batch run ended");
-		if ($batch->has_ready()) {
-			print " <a href='?id=$batch_id&action=restart&batch_id=$batch_id'>Restart batch?</a></td>";
-		} else if ($batch->has_error()) {
-			print " <a href='?id=$batch_id&action=reset&batch_id=$batch_id'>Reset errors?</a>";
-		} else {
-			print " <a href='?action=delete&batch_id=$batch_id'>Delete batch?</a>";
+		if ($owner == $batch->owner) {
+			if ($batch->has_ready()) {
+				print " <a href='?id=$batch_id&action=restart&batch_id=$batch_id'>Restart batch?</a></td>";
+			} else if ($batch->has_error()) {
+				print " <a href='?id=$batch_id&action=reset&batch_id=$batch_id'>Reset errors?</a>";
+			} else {
+				print " <a href='?action=delete&batch_id=$batch_id'>Delete batch?</a>";
+			}
 		}
 	}
